+++ src/cqfantasy/door/__init__.py
@@ -1,0 +1,2 @@
+from .VDoor import VDoor
+from .ArchDoor import ArchDoor